refactor : Move static variable to winnerLottoTest from Lotto
