+++ src/test/java/lotto/domain/LottoGameTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
-import static lotto.domain.LottoTest.winnerLotto;
+import static lotto.domain.WinnerLottoTest.winnerLotto;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
